requirements.txt
aiogram==3.0.0
python-binance==1.0.17
asyncio
